+++ application/services/__init__.py
@@ -1,3 +1,3 @@
-from decision import post_to_decision
+from cases import post_to_cases
 from users import is_matched
 from users import is_owner
