+++ config.py
@@ -4,7 +4,7 @@ class Config(object):
     DEBUG = False
     AUTHENTICATED_SEARCH_API = os.environ['AUTHENTICATED_SEARCH_API']
     SQLALCHEMY_DATABASE_URI = os.environ['DATABASE_URL']
-    DECISION_URL = os.environ['DECISION_URL']
+    CASES_URL = os.environ['CASES_URL']
     SECRET_KEY = os.environ['SECRET_KEY']
     WTF_CSRF_ENABLED = True
     MATCHING_URL = os.environ['MATCHING_URL']
@@ -20,7 +20,7 @@ class Config(object):
 class DevelopmentConfig(Config):
     DEBUG = True
     WTF_CSRF_ENABLED = False
-    DECISION_URL = 'http://decision.landregistry.local'
+    CASES_URL = 'http://cases.landregistry.local'
 
 class TestConfig(DevelopmentConfig):
     TESTING = True
