+++ environment.sh
@@ -4,4 +4,5 @@ export SECRET_KEY='localdev-not-secret'
 export DECISION_URL='http://decision.landregistry.local'
 export MATCHING_URL='http://matching.landregistry.local'
 export OWNERSHIP_URL='http://ownership.landregistry.local'
+export CASES_URL='http://cases.landregistry.local'
 export OS_API_KEY='no-key'
