+++ environment_test.sh
@@ -4,4 +4,5 @@ export SECRET_KEY='localdev-not-secret'
 export DECISION_URL='http://nowhere'
 export MATCHING_URL='http://nowhere'
 export OWNERSHIP_URL='http://nowhere'
+export CASES_URL='http://nowhere'
 export OS_API_KEY='no-key-in-test'
